+++ src/main/java/com/urlshortener/url/component/ShortUrlGenerator.java
@@ -1,4 +1,4 @@
-package com.urlshortener.shorturl.component;
+package com.urlshortener.url.component;
 
 public interface ShortUrlGenerator {
     String generate(String source);
