+++ src/main/java/com/urlshortener/url/model/dto/CreateResponse.java
@@ -1,4 +1,4 @@
-package com.urlshortener.shorturl.model.dto;
+package com.urlshortener.url.model.dto;
 
 import lombok.Data;
