+++ src/main/resources/application.yaml
@@ -23,3 +23,5 @@ spring:
       minimum-idle: 1
       maximum-pool-size: 5
       pool-name: H2_DB
+server:
+  port: 5000
